mitigate reboot first time because of raspi no clock

+++ system/watchdog.c
@@ -98,6 +98,7 @@ void wd_action()
 {
     int i;
     unsigned long lcur_time;
+	static int reboot_cnt = 0;
     lcur_time = get_sys_cur_time();
     for(i=0; i < UNKNOWN_MODULE_ID; i++)
     {
@@ -113,7 +114,13 @@ void wd_action()
 		{
 	        logging(DBG_ERROR, "System about to be reboot because %s mtime: %lu ctime: %lu\n", 
 					modname[modulelist[i].module_id], modulelist[i].timer, lcur_time);
-			system("reboot");
+			reboot_cnt++;	// Because raspi had no clock so a lot of time first boot after long
+							// blackout will cause it reboot
+			if(reboot_cnt > 3)
+				system("reboot");
+			else
+				reboot_cnt++;	// Because raspi had no clock so a lot of time first boot after long
+								// blackout will cause it reboot
 		}
     }
 }
